apps/api/searching-svc/pyproject.toml
[tool.coverage.run]
branch = true
source = ["searching_svc"]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'../../../coverage/apps/api/searching-svc/html' --cov-report xml:'../../../coverage/apps/api/searching-svc/coverage.xml' --html='../../../reports/apps/api/searching-svc/unittests/html/index.html' --junitxml='../../../reports/apps/api/searching-svc/unittests/junit.xml'"

[tool.poetry]
name = "searching-svc"
version = "1.0.0"
description = "Searching service"
authors = ['Hutech University <hutech@hutech.edu.vn>']
license = 'MIT'

[[tool.poetry.packages]]
include = "searching_svc"

[tool.poetry.dependencies]
python = ">=3.9 <=3.12"
grpcio = "^1.59.3"
grpcio-tools = "^1.59.3"
pydantic = "1.10.0"
protobuf = "^4.25.1"
python-dotenv = "^1.0.0"
farm-haystack = {extras = ["elasticsearch"], version = "^1.22.1"}

[tool.poetry.group.dev.dependencies]
autopep8 = "2.0.2"
flake8 = "6.0.0"
pytest = "7.3.1"
pytest-sugar = "0.9.7"
pytest-cov = "4.1.0"
pytest-html = "3.2.0"
isort = "^5.12.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[tool.nx]
autoActivate = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
